+++ src/text_selection_tests/kld/kld_iterator_py/test_get_minimum_kld_keys.py
@@ -4,9 +4,9 @@ from text_selection.kld.kld_iterator import get_minimum_kld_keys
 
 
 def test_componenttest_without_preselection():
-  data = np.array([[1, 2], [1, 1], [1, 1], [1, 1], [0, 0]], dtype=np.uint32)
+  data = np.array([[1, 2], [1, 1], [1, 1], [1, 1], [0, 0]], dtype=np.uint16)
   keys = OrderedSet((0, 1, 3, 4))
-  covered = np.array([0, 0], dtype=np.uint32)
+  covered = np.array([0, 0], dtype=np.uint16)
   target_dist = np.full(shape=(5, 2), fill_value=0.5, dtype=np.float64)
   div, result = get_minimum_kld_keys(
     data=data,
@@ -20,9 +20,9 @@ def test_componenttest_without_preselection():
 
 
 def test_componenttest_with_preselection():
-  data = np.array([[1, 2], [2, 1], [2, 1], [2, 1], [0, 0]], dtype=np.uint32)
+  data = np.array([[1, 2], [2, 1], [2, 1], [2, 1], [0, 0]], dtype=np.uint16)
   keys = OrderedSet((0, 1, 3, 4))
-  covered = np.array([1, 2], dtype=np.uint32)
+  covered = np.array([1, 2], dtype=np.uint16)
   target_dist = np.full(shape=(5, 2), fill_value=0.5, dtype=np.float64)
   div, result = get_minimum_kld_keys(
     data=data,
@@ -36,9 +36,9 @@ def test_componenttest_with_preselection():
 
 
 def test_componenttest_with_preselection_zero_is_selected():
-  data = np.array([[1, 2], [2, 1], [2, 1], [2, 1], [0, 0]], dtype=np.uint32)
+  data = np.array([[1, 2], [2, 1], [2, 1], [2, 1], [0, 0]], dtype=np.uint16)
   keys = OrderedSet((0, 1, 3, 4))
-  covered = np.array([1, 1], dtype=np.uint32)
+  covered = np.array([1, 1], dtype=np.uint16)
   target_dist = np.full(shape=(5, 2), fill_value=0.5, dtype=np.float64)
   div, result = get_minimum_kld_keys(
     data=data,
